requirements.txt
fastapi
sqlalchemy
psycopg2
pydantic
python-dotenv 
uvicorn
python-jose
PyMySQL
passlib
jwt
cryptography
bcrypt